修复: 精排输入 tensor 类型由 float32 改为 int64
bge-reranker-v2-m3 ONNX 模型要求 input_ids / attention_mask /
token_type_ids 为 int64 输入，原先用 float[][] 创建 tensor 会导致
推理失败。token 数据本就是 long[]，改为 long[][] 直接赋值无需转换。

+++ src/main/java/me/maxt/rag/web/service/vector/rerank/CrossEncoderReranker.java
@@ -83,9 +83,10 @@ public class CrossEncoderReranker implements Reranker {
         }
 
         int n = candidates.size();
-        float[][] inputIds = new float[n][MAX_SEQ_LENGTH];
-        float[][] attentionMask = new float[n][MAX_SEQ_LENGTH];
-        float[][] tokenTypeIds = new float[n][MAX_SEQ_LENGTH];
+        // bge-reranker-v2-m3 的 token 输入是 int64，需用 long[][]（float32 会导致推理失败）
+        long[][] inputIds = new long[n][MAX_SEQ_LENGTH];
+        long[][] attentionMask = new long[n][MAX_SEQ_LENGTH];
+        long[][] tokenTypeIds = new long[n][MAX_SEQ_LENGTH];
 
         for (int i = 0; i < n; i++) {
             String passage = candidates.get(i).embedded().text();
